FMonapresBE/requirements.txt
affine==2.4.0
asgiref==3.6.0
attrs==23.1.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
Django==4.2
djangorestframework==3.14.0
Fiona==1.9.3
geopandas==0.12.2
importlib-metadata==6.6.0
munch==2.5.0
numpy==1.24.3
packaging==23.1
pandas==2.0.1
pyparsing==3.0.9
pyproj==3.5.0
python-dateutil==2.8.2
pytz==2023.3
rasterio==1.3.6
shapely==2.0.1
six==1.16.0
snuggs==1.4.7
sqlparse==0.4.4
tzdata==2023.3
zipp==3.15.0
